Move peer behaviour package to lib
